+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Initital project setup APPS-569
 * Add Example APPS-581
 * Network Layer APPS-588
-* Phrase Integration APPS APPS-583
+* Phrase Integration APPS-583
+* pointfreeco/swift-tagged 0.9.0
 
 ### Updated
 * realm/SwiftLint 0.50.3 (was 0.50.0)